django/0519/자료수집예제/requirements.txt
asgiref==3.3.4
certifi==2021.10.8
charset-normalizer==2.0.7
Django==3.2.3
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
idna==3.3
PyJWT==1.7.1
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.7
